+++ addons/confirma/Plugin.cs
@@ -1,4 +1,5 @@
 #if TOOLS
+using Confirma.Helpers;
 using Godot;
 
 namespace Confirma;
@@ -13,11 +14,18 @@ public partial class Plugin : EditorPlugin
     public override void _EnterTree()
     {
         _testBottomPanel = GD.Load<PackedScene>(
-            GetPluginPath() + "/src/scenes/confirma_bottom_panel/ConfirmaBottomPanel.tscn"
+            GetPluginPath()
+            + "/src/scenes/confirma_bottom_panel/ConfirmaBottomPanel.tscn"
         ).Instantiate<Control>();
         _ = AddControlToBottomPanel(_testBottomPanel, "Confirma");
 
-        AddAutoloadSingleton("Confirma", GetPluginPath() + "/src/scenes/confirma_autoload/ConfirmaAutoload.tscn");
+        AddAutoloadSingleton(
+            "Confirma",
+            GetPluginPath()
+            + "/src/scenes/confirma_autoload/ConfirmaAutoload.tscn"
+        );
+
+        SetUpSettings();
 
         GD.Print("Confirma is ready!");
     }
@@ -35,6 +43,28 @@ public partial class Plugin : EditorPlugin
     private string GetPluginPath()
     {
         return GetScript().As<Script>().ResourcePath.GetBaseDir();
+    }
+
+    public static string GetPluginLocation()
+    {
+        // Read the information from the project settings
+        // because addons can be installed in various locations,
+        // so it cannot be assumed that it will always be in the default location.
+        string p = ProjectSettings.GetSetting("autoload/Confirma").AsString();
+        return p[1..].Remove(p.Length - 51);
+    }
+
+    public static void SetUpSettings()
+    {
+        // Note: When changing path here,
+        // remember to change it also in TestsProps.cs.
+        _ = Settings.CreateSetting(
+            "confirma/config/gdscript_tests_folder",
+            "res://gdtests/",
+            "res://gdtests/",
+            true,
+            true
+        );
     }
 }
 #endif