+++ addons/confirma/src/classes/TestLog.cs
@@ -10,6 +10,7 @@ public class TestLog
     public string? Name { get; }
     public ETestCaseState State { get; } = ETestCaseState.Ignored;
     public ELogType Type { get; }
+    public ELangType Lang { get; set; } = ELangType.None;
 
     public TestLog(ELogType type)
     {
@@ -22,12 +23,20 @@ public class TestLog
         Message = message;
     }
 
+    public TestLog(ELogType type, ELangType testLang ,string message)
+    {
+        Type = type;
+        Message = message;
+        Lang = testLang;
+    }
+
     public TestLog(
         ELogType type,
         string name,
         ETestCaseState state,
         string parameters = "",
-        string? message = null
+        string? message = null,
+        ELangType lang = ELangType.None
     )
     {
         Type = type;
@@ -38,6 +47,7 @@ public class TestLog
             : string.Empty
         );
         State = state;
+        Lang = lang;
     }
 
     public static string GetTestCaseStateString(ETestCaseState state)
@@ -74,7 +84,7 @@ public class TestLog
                 }
                 break;
             case ELogType.Class:
-                Log.Print($"> {Colors.ColorText(Message!, Colors.Class)}...");
+                Log.Print($"> {GetLangHeader()} {Colors.ColorText(Message!, Colors.Class)}...");
                 break;
             case ELogType.Info:
                 if (Message != null)
@@ -104,8 +114,9 @@ public class TestLog
     {
         string color = GetTestCaseStateColor(State);
         string sState = GetTestCaseStateString(State);
+        string langColor = getLangColor();
 
-        Log.PrintLine($"| {Name}... ");
+        Log.PrintLine($"{Colors.ColorText("|",langColor)} {Name}... ");
 
         if (State != ETestCaseState.Passed)
         {
@@ -115,7 +126,7 @@ public class TestLog
 
         if (Message is not null)
         {
-            Log.PrintLine($"  |- {Colors.ColorText(Message, color)}");
+            Log.PrintLine($"{Colors.ColorText("|",langColor)}- {Colors.ColorText(Message, color)}");
         }
     }
 
@@ -123,12 +134,43 @@ public class TestLog
     {
         string color = GetTestCaseStateColor(State);
         string sState = GetTestCaseStateString(State);
+        string langColor = getLangColor();
 
-        Log.PrintLine($"| {Name}... {Colors.ColorText(sState, color)}.");
+        Log.PrintLine($" {Colors.ColorText("|",langColor)} {Name}... {Colors.ColorText(sState, color)}.");
 
         if (Message is not null)
         {
-            Log.PrintLine($"- {Colors.ColorText(Message, color)}");
+            Log.PrintLine($"{Colors.ColorText("|",langColor)}- {Colors.ColorText(Message, color)}");
         }
+    }
+
+    private string GetLangHeader ()
+    {
+        string langText = "N/A";
+        string color = getLangColor();
+
+        switch (Lang)
+        {
+            case ELangType.CSharp:
+                langText = "C#";
+                break;
+            case ELangType.GDScript:
+                langText = "GDScript";
+                break;
+        }
+
+        return Colors.ColorText($"[{langText}]",color);
+    }
+
+    private string getLangColor ()
+    {
+        switch (Lang)
+        {
+            case ELangType.CSharp:
+                return Colors.CSharp;
+            case ELangType.GDScript:
+                return Colors.Gdscript;
+        }
+        return Colors.Error;
     }
 }