+++ addons/confirma/src/gd_assertions/confirm_array.gd
@@ -1,51 +1,36 @@
+@tool
 class_name ConfirmArray
+
+static var exts: CSharpScript = load(
+	GdHelper.get_plugin_path() + "wrappers/ConfirmArrayWrapper.cs"
+).new()
 
 
 static func is_of_size(actual: Array, expected: int, message: String = "") -> Array:
-	if (not actual.size() == expected):
-		Confirma.emit_signal(
-			"GdAssertionFailed", 
-			"Array size is %s, but expected %s." % actual.size(), expected \
-			if message.is_empty() else message
-		)
+	exts.ConfirmSize(actual, expected, message)
+
 	return actual
 
 
 static func is_empty(actual: Array, message: String = "") -> Array:
-	if (not actual.size() == 0):
-		Confirma.emit_signal(
-			"GdAssertionFailed",
-			"Array is not empty." \
-			if message.is_empty() else message
-		)
+	exts.ConfirmEmpty(actual, message)
+
 	return actual
 
 
 static func is_not_empty(actual: Array, message: String = "") -> Array:
-	if (not actual.size() > 0):
-		Confirma.emit_signal(
-			"GdAssertionFailed",
-			"Array is empty." \
-			if message.is_empty() else message
-		)
+	exts.ConfirmNotEmpty(actual, message)
+
 	return actual
 
 
 static func contains(actual: Array, expected: Variant, message: String = "") -> Array:
-	if (not actual.has(expected)):
-		Confirma.emit_signal(
-			"GdAssertionFailed",
-			"Array does not contain '%s'." % expected \
-			if message.is_empty() else message
-		)
+	exts.ConfirmContains(actual, expected, message)
+
 	return actual
 
 
 static func not_contains(actual: Array, expected: Variant, message: String = "") -> Array:
-	if (actual.has(expected)):
-		Confirma.emit_signal(
-			"GdAssertionFailed",
-			"Array contains '%s'." % expected \
-			if message.is_empty() else message
-		)
+	exts.ConfirmNotContains(actual, expected, message)
+
 	return actual