+++ addons/confirma/src/helpers/Colors.cs
@@ -8,7 +8,9 @@ public static class Colors
     public static readonly string Success = "#8eef97";
     public static readonly string Warning = "#ffdd65";
     public static readonly string Error = "#ff786b";
-    public static readonly string Class = "#22A7EB";
+    public static readonly string Class = "#218a72";
+    public static readonly string CSharp = "#9b4993";
+    public static readonly string Gdscript = "#478cbf";
     public static readonly string TerminalReset = "\x1b[0m";
 
     public static string ColorText<T>(T text, string color)