+++ addons/confirma/src/scenes/confirma_autoload/ConfirmaAutoload.cs
@@ -12,12 +12,9 @@ namespace Confirma.Scenes;
 [Tool]
 public partial class ConfirmaAutoload : Node
 {
-    [Signal]
-    public delegate void GdAssertionFailedEventHandler(string message);
-
     public TestsProps Props = new();
 
-    private bool _usedConfirmaApi = false;
+    private bool _usedConfirmaApi;
 
     public override void _Ready()
     {
@@ -86,7 +83,7 @@ public partial class ConfirmaAutoload : Node
                         "Invalid value: argument '--confirma-run' cannot be empty"
                         + " when using argument '--confirma-method'.\n"
                     );
-                    return false; ;
+                    return false;
                 }
 
                 string method = ParseArgumentContent(arg);