+++ addons/confirma/src/scenes/confirma_bottom_panel/ConfirmaBottomPanel.cs
@@ -1,5 +1,6 @@
 #if TOOLS
 
+using Confirma.Enums;
 using Godot;
 
 namespace Confirma.Scenes;
@@ -8,9 +9,9 @@ namespace Confirma.Scenes;
 public partial class ConfirmaBottomPanel : Control
 {
 #nullable disable
-    private Button _runAllTests, _runCSharpTests, _runGdScriptTests, _clearOutput;
-    private CheckBox _verbose;
+    private Button _runAllTests, _runCSharpTests, _runGdScriptTests, _clearOutput, _settings;
     private TestRunnerEditor _testRunner;
+    private Window _settingsWindow;
     private ConfirmaAutoload _autoload;
 #nullable restore
 
@@ -28,20 +29,21 @@ public partial class ConfirmaBottomPanel : Control
         _clearOutput = GetNode<Button>("%ClearOutput");
         _clearOutput.Pressed += OnClearOutputPressed;
 
-        _verbose = GetNode<CheckBox>("%Verbose");
-
         _testRunner = GetNode<TestRunnerEditor>("%TestRunnerEditor");
         _testRunner.TestsExecutionStarted += OnTestsExecutionStarted;
         _testRunner.TestsExecutionFinished += OnTestsExecutionFinished;
 
+        _settings = GetNode<Button> ("%Settings");
+        _settings.Pressed += OnSettingsPressed;
+
+        _settingsWindow = GetNode<Window> ("%SettingsWindow");
+
         _ = CallDeferred("LateInit");
     }
 
-    private void LateInit()
+        private void LateInit()
     {
         _autoload = GetNode<ConfirmaAutoload>("/root/Confirma");
-
-        _verbose.Toggled += (bool on) => _autoload.Props.IsVerbose = on;
     }
 
     private void ResetLanguagesToggle()
@@ -75,6 +77,11 @@ public partial class ConfirmaBottomPanel : Control
     private void OnClearOutputPressed()
     {
         _testRunner.ClearOutput();
+    }
+
+    private void OnSettingsPressed()
+    {
+        _settingsWindow.Show();
     }
 
     private void OnTestsExecutionStarted()