+++ gdtests/wrappers/confirm_array_wrapper_test.gd
@@ -1,4 +1,4 @@
-class_name ConfirmArrayTest
+class_name ConfirmArrayWrapperTest extends TestClass
 
 
 func confirm_is_of_size_when_is_of_size() -> void: