+++ gdtests/wrappers/confirm_signal_wrapper_test.gd
@@ -1,0 +1,16 @@
+class_name ConfirmSignalWrapperTest extends TestClass
+
+func confirm_signal_exists_when_exists() -> void:
+	var btn := Button.new()
+
+	ConfirmSignal.exists(btn, "pressed")
+
+	btn.free()
+
+
+func confirm_signal_does_not_exist_when_does_not_exist() -> void:
+	var btn := Button.new()
+
+	ConfirmSignal.does_not_exist(btn, "pressed_heavily")
+
+	btn.free()